+++ packages/polythene-core-toolbar/src/vars.js
@@ -24,7 +24,6 @@ export default {
   color_light_text:          rgba(vars.color_light_foreground, vars.blend_light_text_primary),
   color_dark_text:           rgba(vars.color_dark_foreground,  vars.blend_dark_text_primary),
 
-  // background colors may be set in theme; disabled by default
-  // color_light_background:    "transparent",
-  // color_dark_background:     "transparent",
+  color_light_background:    rgba(vars.color_light_background),
+  color_dark_background:     rgba(vars.color_dark_background)
 };